+++ core/src/main/scala/org/locationtech/rasterframes/extensions/DataFrameMethods.scala
@@ -23,6 +23,7 @@ package org.locationtech.rasterframes.extensions
 
 import geotrellis.proj4.CRS
 import geotrellis.layer._
+import geotrellis.raster.resample.{NearestNeighbor, ResampleMethod => GTResampleMethod}
 import geotrellis.util.MethodExtensions
 import geotrellis.vector.Extent
 import org.apache.spark.sql.catalyst.expressions.Attribute
@@ -37,6 +38,7 @@ import org.locationtech.rasterframes.util._
 import org.locationtech.rasterframes.{MetadataKeys, RasterFrameLayer}
 import spray.json.JsonFormat
 import org.locationtech.rasterframes.util.JsonCodecs._
+
 import scala.util.Try
 
 /**
@@ -165,9 +167,10 @@ trait DataFrameMethods[DF <: DataFrame] extends MethodExtensions[DF] with Metada
     * }}}
     *
     * @param right Right side of the join.
+    * @param resampleMethod string indicating method to use for resampling.
     * @return joined dataframe
     */
-  def rasterJoin(right: DataFrame): DataFrame = RasterJoin(self, right, None)
+  def rasterJoin(right: DataFrame, resampleMethod: GTResampleMethod = NearestNeighbor): DataFrame = RasterJoin(self, right, resampleMethod, None)
 
   /**
     * Performs a jeft join on the dataframe `right` to this one, reprojecting and merging tiles as necessary.
@@ -183,10 +186,11 @@ trait DataFrameMethods[DF <: DataFrame] extends MethodExtensions[DF] with Metada
     * @param leftCRS this (left) datafrasme's CRS column
     * @param rightExtent right dataframe's CRS extent
     * @param rightCRS right dataframe's CRS column
+    * @param resampleMethod string indicating method to use for resampling.
     * @return joined dataframe
     */
-  def rasterJoin(right: DataFrame, leftExtent: Column, leftCRS: Column, rightExtent: Column, rightCRS: Column): DataFrame =
-    RasterJoin(self, right, leftExtent, leftCRS, rightExtent, rightCRS, None)
+  def rasterJoin(right: DataFrame, leftExtent: Column, leftCRS: Column, rightExtent: Column, rightCRS: Column, resampleMethod: GTResampleMethod): DataFrame =
+    RasterJoin(self, right, leftExtent, leftCRS, rightExtent, rightCRS, resampleMethod, None)
 
   /**
     * Performs a jeft join on the dataframe `right` to this one, reprojecting and merging tiles as necessary.
@@ -200,10 +204,11 @@ trait DataFrameMethods[DF <: DataFrame] extends MethodExtensions[DF] with Metada
     * @param leftCRS this (left) datafrasme's CRS column
     * @param rightExtent right dataframe's CRS extent
     * @param rightCRS right dataframe's CRS column
+    * @param resampleMethod string indicating method to use for resampling.
     * @return joined dataframe
     */
-  def rasterJoin(right: DataFrame, joinExpr: Column, leftExtent: Column, leftCRS: Column, rightExtent: Column, rightCRS: Column): DataFrame =
-    RasterJoin(self, right, joinExpr, leftExtent, leftCRS, rightExtent, rightCRS, None)
+  def rasterJoin(right: DataFrame, joinExpr: Column, leftExtent: Column, leftCRS: Column, rightExtent: Column, rightCRS: Column, resampleMethod: GTResampleMethod): DataFrame =
+    RasterJoin(self, right, joinExpr, leftExtent, leftCRS, rightExtent, rightCRS, resampleMethod, None)
 
 
   /** Layout contents of RasterFrame to a layer. Assumes CRS and extent columns exist. */