+++ pyrasterframes/src/main/python/pyrasterframes/rf_ipython.py
@@ -17,6 +17,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 #
+from functools import partial
 
 import pyrasterframes.rf_types
 from pyrasterframes.rf_types import Tile
@@ -24,7 +25,7 @@ from shapely.geometry.base import BaseGeometry
 from matplotlib.axes import Axes
 import numpy as np
 from pandas import DataFrame
-from typing import Optional, Tuple
+from typing import Optional, Tuple, Union
 
 _png_header = bytearray([0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A])
 
@@ -50,7 +51,6 @@ def plot_tile(tile: Tile, normalize: bool = True, lower_percentile: float = 1., 
     -------
     created or modified axis object
     """
-
     if axis is None:
         import matplotlib.pyplot as plt
         axis = plt.gca()
@@ -58,7 +58,8 @@ def plot_tile(tile: Tile, normalize: bool = True, lower_percentile: float = 1., 
     arr = tile.cells
 
     def normalize_cells(cells: np.ndarray) -> np.ndarray:
-        assert upper_percentile > lower_percentile, 'invalid upper and lower percentiles {}, {}'.format(lower_percentile, upper_percentile)
+        assert upper_percentile > lower_percentile, 'invalid upper and lower percentiles {}, {}'.format(
+            lower_percentile, upper_percentile)
         sans_mask = np.array(cells)
         lower = np.nanpercentile(sans_mask, lower_percentile)
         upper = np.nanpercentile(sans_mask, upper_percentile)
@@ -90,7 +91,7 @@ def tile_to_png(tile: Tile, lower_percentile: float = 1., upper_percentile: floa
     from matplotlib.figure import Figure
 
     # Set up matplotlib objects
-    nominal_size = 3  # approx full size for a 256x256 tile
+    nominal_size = 2
     if fig_size is None:
         fig_size = (nominal_size, nominal_size)
 
@@ -105,9 +106,9 @@ def tile_to_png(tile: Tile, lower_percentile: float = 1., upper_percentile: floa
 
     if title is None:
         axis.set_title('{}, {}'.format(tile.dimensions(), tile.cell_type.__repr__()),
-                       fontsize=fig_size[0]*4)  # compact metadata as title
+                       fontsize=fig_size[0] * 4)  # compact metadata as title
     else:
-        axis.set_title(title, fontsize=fig_size[0]*4)  # compact metadata as title
+        axis.set_title(title, fontsize=fig_size[0] * 4)  # compact metadata as title
 
     with io.BytesIO() as output:
         canvas.print_png(output)
@@ -123,7 +124,7 @@ def tile_to_html(tile: Tile, fig_size: Optional[Tuple[int, int]] = None) -> str:
     return b64_img_html.format(b64_png)
 
 
-def binary_to_html(blob):
+def binary_to_html(blob) -> Union[str, bytearray]:
     """ When using rf_render_png, the result from the JVM is a byte string with special PNG header
         Look for this header and return base64 encoded HTML for Jupyter display
     """
@@ -136,7 +137,7 @@ def binary_to_html(blob):
         return blob
 
 
-def pandas_df_to_html(df: DataFrame) -> str:
+def pandas_df_to_html(df: DataFrame) -> Optional[str]:
     """Provide HTML formatting for pandas.DataFrame with rf_types.Tile in the columns.  """
     import pandas as pd
     # honor the existing options on display
@@ -170,9 +171,9 @@ def pandas_df_to_html(df: DataFrame) -> str:
         if isinstance(g, BaseGeometry):
             wkt = g.wkt
             if len(wkt) > default_max_colwidth:
-                return wkt[:default_max_colwidth-3] + '...'
+                return wkt[:default_max_colwidth - 3] + '...'
             else:
-                wkt
+                return wkt
         else:
             return g.__repr__()
 
@@ -188,7 +189,7 @@ def pandas_df_to_html(df: DataFrame) -> str:
     formatter.update({c: _safe_bytearray_to_html for c in bytearray_cols})
 
     # This is needed to avoid our tile being rendered as `<img src="only up to fifty char...`
-    pd.set_option('display.max_colwidth', -1)
+    pd.set_option('display.max_colwidth', None)
     return_html = df.to_html(escape=False,  # means our `< img` does not get changed to `&lt; img`
                              formatters=formatter,  # apply custom format to columns
                              render_links=True,  # common in raster frames
@@ -213,55 +214,75 @@ def spark_df_to_html(df: DataFrame, num_rows: int = 5, truncate: bool = False) -
 
 def _folium_map_formatter(map) -> str:
     """ inputs a folium.Map object and returns html of rendered map """
-    
+
     import base64
     html_source = map.get_root().render()
     b64_source = base64.b64encode(
         bytes(html_source.encode('utf-8'))
-        ).decode('utf-8')
+    ).decode('utf-8')
 
-    source_blob = '<iframe src="data:text/html;charset=utf-8;base64,{}" allowfullscreen="" webkitallowfullscreen="" mozallowfullscreen="" style="position:relative;width:100%;height:500px"></iframe>' 
+    source_blob = '<iframe src="data:text/html;charset=utf-8;base64,{}" allowfullscreen="" webkitallowfullscreen="" mozallowfullscreen="" style="position:relative;width:100%;height:500px"></iframe>'
     return source_blob.format(b64_source)
 
 
 try:
     from IPython import get_ipython
-    from IPython.display import display_png, display_markdown, display
+    from IPython.display import display_png, display_markdown, display_html, display
+
     # modifications to currently running ipython session, if we are in one; these enable nicer visualization for Pandas
     if get_ipython() is not None:
         import pandas
         import pyspark.sql
         from pyrasterframes.rf_types import Tile
-        ip = get_ipython()
 
+        ip = get_ipython()
+        formatters = ip.display_formatter.formatters
         # Register custom formatters
-        html_formatter = ip.display_formatter.formatters['text/html']
+        # PNG
+        png_formatter = formatters['image/png']
+        png_formatter.for_type(Tile, tile_to_png)
+        # HTML
+        html_formatter = formatters['text/html']
         html_formatter.for_type(pandas.DataFrame, pandas_df_to_html)
         html_formatter.for_type(pyspark.sql.DataFrame, spark_df_to_html)
+        html_formatter.for_type(Tile, tile_to_html)
 
-        # these will likely only effect docs build
-        markdown_formatter = ip.display_formatter.formatters['text/markdown']
+        # Markdown. These will likely only effect docs build.
+        markdown_formatter = formatters['text/markdown']
+        # Pandas doesn't have a markdown
+        markdown_formatter.for_type(pandas.DataFrame, pandas_df_to_html)
         markdown_formatter.for_type(pyspark.sql.DataFrame, spark_df_to_markdown)
+        # Running loose here by embedding tile as `img` tag.
+        markdown_formatter.for_type(Tile, tile_to_html)
 
         try:
             # this block is to try to avoid making an install dep on folium but support if in the environment
             import folium
+
             markdown_formatter.for_type(folium.Map, _folium_map_formatter)
         except ImportError as e:
             pass
 
-        png_formatter = ip.display_formatter.formatters['image/png']
-        png_formatter.for_type(Tile, tile_to_png)
-
-        # These are done for those few cases where we need to set the number of rows and/or truncate option
-        # Can be removed if we can figure out a way to pass settings through `display`
-        pyspark.sql.DataFrame.showMarkdown = spark_df_to_markdown
-        pyspark.sql.DataFrame.showHTML = spark_df_to_html
         Tile.show = plot_tile
 
-        # This is a trick we may have to introduce above.
-        # from IPython.display import display_html
-        # display_html(rf.showHTML(truncate=True), raw=True)
+        # noinspection PyTypeChecker
+        def _display(df: pyspark.sql.DataFrame, num_rows: int = 5, truncate: bool = False,
+                     mimetype: str = 'text/html') -> ():
+            """
+            Invoke IPython `display` with specific controls.
+            :param num_rows: number of rows to render
+            :param truncate: If `True`, shorten width of columns to no more than 40 characters
+            :return: None
+            """
+
+            if "html" in mimetype:
+                display_html(spark_df_to_html(df, num_rows, truncate), raw=True)
+            else:
+                display_markdown(spark_df_to_markdown(df, num_rows, truncate), raw=True)
+
+
+        # Add enhanced display function
+        pyspark.sql.DataFrame.display = _display
 
 except ImportError as e:
     pass